Rename OpConfig Field for Readability (#3116)
### Ticket
#2441 

### Problem description
#2979 addressed the functionality required for #2441, but introduced
awkward and difficult to understand statements like `opConfig.config`.

The `OpConfig` struct represents a particular configuration of a TTNN Op
for the purpose of the optimizer. It has two fields:
- `TTNNLayoutAttr outputLayout` representing a specific output tensor
layout for the op
- `Attribute config` representing op-specific configuration structs
(e.g. matmul or conv2d)

We should better distinguish between `OpConfig` and its field `config`.

### What's changed
- Rename `config` to `opSpecificAttr` so `opConfig.config` becomes
`opConfig.opSpecificAttr`

#### Details
The initial discussion in #2979 was centred around renaming `OpConfig`
to `OpParams` or `TTNNOperationParams`. However, this would necessitate
changing a large number of variables and APIs within the optimizer to
match, without significantly improving readability.

`opParams.config` is only slightly better than `opConfig.config`;
however, the optimizer has many APIs in the form of `getXConfig` and
many variables and containers in the form `xConfigs` all dealing with
`OpConfig` objects. These APIs and variables would all need to be
renamed to `params` to match `OpParams`.

The new field name is now more clear that this attribute is op-specific

### Checklist
- [X] New/Existing tests provide coverage for changes