TTIR elementwise ops signature splitted by arity (#3102)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2829

### Problem description
TTIR elementwise ops provided an interface that turned out to be too
generic for the majority of use-cases. We paid the price of that form,
but never reaped any benefit from it (the only one that I could find is
that lowering to Linalg is a little bit easier).

### What's changed
Changed the ops interface, such that they have a concrete signature that
depends on op arity.

### Checklist
- [x] New/Existing tests provide coverage for changes