TTIR pass to convert quantized integer types to i32 (#2862)
### Ticket
Required for https://github.com/tenstorrent/tt-mlir/issues/2464

### Problem description
Currently the runtime implementation of quant/dequant/requant can only
handle int32 integer types.

### What's changed
A pass --ttir-quant-data-type-conversion has been introduced which takes
in a quant-data-type (e.g: 'quant-data-type=int16') and performs the
conversion. The default is set to i32.
 
A TTNN workaround was considered but the mechanism to modify the op
(toLayout) was not appropriate in this case; there is no concept of a
quantized data type for TTNN layouts.

### Checklist
- [ ] New transform test for TTIR has been added for
quant/dequant/requant.