Fix const-eval bug for case where subgraphs only share creation op (#3175)
### Ticket
NA

### Problem description
@AleksKnezevic found a bug in const-eval during tt-torch testing. In the
case of `CreationOps`, we want to treat them specially s.t. not ever
tensor creation is always const-eval'ed. However, I was missing logic to
properly treat the Values of a creationOp which has been merged into a
subgraph as part of that subgraph w.r.t. merging other subgraphs.

### What's changed
Any creationOp added to a given subgraph also has the values it produces
treated as part of this subgraph, s.t. proper merging occurs when
multiple subgraphs share a creationOp

### Checklist
- [x] New/Existing tests provide coverage for changes