Add workaround for int broadcasting in 'Add' op (#3164)
### Ticket
[#489](https://github.com/tenstorrent/tt-torch/issues/489) in tt-torch

### Problem description
TTNN currently fails to handle broadcasting for int32 inputs in the
`add` op, which is used internally to represent `expand` operations.
When an `expand` op with int32 is lowered to `add` with broadcasting,
execution does not work as expected.

### What's changed
As a temporary workaround, inputs are converted to bfloat16 when shapes
differ. This should be removed once TTNN properly supports broadcasting
for int.

### Checklist
- [x] full evaluation of bert model