Fusing (#3099)
Adding basic fusing support:
	•	`conv2d + add` → fused into `conv2d` with bias (in **TTIR**)
	•	`conv2d + relu` → fused into `conv2d` with activation (in **TTNN**)
	•	`reduction + reshape` → fused into reduction (in **TTIR**)
	•	`exp(x) / sum(exp(x))` → fused into `softmax` (in **TTIR**)
• `conv2d + multiply` → commuted into `reshape + multiply + conv2d` (in
**TTIR**)
_(only applied when multiply can be constant-evaluated)_

Other possible fusions (future work):
• `linear / conv2d_transpose + add` → fused into `linear /
conv2d_transpose` with bias (in TTIR)
	•	`matmul + activation` → fused into `matmul` with activation (in TTNN)

Additional fusion opportunities involving commuting operations through
the graph are possible but will initially be handled case-by-case.

Note:
This fusion will be integrated into the pipeline later. For now, it will
likely be inserted at the beginning of the passes for the **TTIR** and
**TTNN** dialects.