Add workaround for concat op (#3188)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3193

### Problem description
Concat op is failing for following two cases. 
1. If all input tensors have shape of [1, 1] or [1, 1, 1] and concat is
performed on last dimension then the generated output is incorrect. The
first element of the output is correct while all other elements are
zero.
2. The second case is extension of the first one if any tensor's last
dim is not `1` (but not for all tensors) then it crashes with `TT_FATAL`
error. For example we want to concatenate 3 tensors and the shapes are
[1, 32], [1, 2], [1, 1] then this will crash.

### What's changed
Add one additional dimension to all the input tensor (using reshape) for
the cases described above so that concatenation is not applied on the
last dim of the input tensor. After concatenation; reshape backe the
output tensor to original shape.

### Checklist
- [X] New/Existing tests provide coverage for changes