Make ttkernel traits more precise (#3187)
### Ticket
Closes #3186
Closes #3010

### Problem description
The `ComputeOpsRewriter` is getting longer and longer, we should split
it into FPU & SFPU rewriters.

### What's changed
Create FPU & SFPU rewriters that handles different types of ops.
The remaining tiling handling parts are now in the
`TTIRTilizeUntilizeRewriter`.
Traits for different arities are also created and assigned to the
corresponding ops. The rewriters will generate init ops accordingly.
Also add support for the sine SFPU op to validate the change.

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Signed-off-by: wenbinlyuTT <wenbinlyu@tenstorrent.com>