Workaround for max and min reductions when dimensions are not multiple of 32 (#3246)
### Ticket


### Problem description
There is a bug in metal (
https://github.com/tenstorrent/tt-metal/issues/21615 ) where max
reduction with all negative inputs erroneously returns zero instead of
the real maximum in cases where the dimension is not divisible by 32.
This is hit in some model tests from tt-xla.
Same failure mode is observed for min and all positive inputs.

### What's changed
I added a TTNN workaround to pad the input tensor with the neutral
element for reduction in cases where the input has reduction dimensions
that aren't a multiple of 32.

### Checklist
- [X] New/Existing tests provide coverage for changes - I modified most
of the tests to not hit the workaround so it is clearer what is being
tested and added workaround specific tests.