Swap personal in favor of team aliases (#3160)
### Ticket
N/A

### Problem description
N/A

### What's changed
@staylorTT has added [mlir
teams](https://github.com/orgs/tenstorrent/teams/forge/teams) within the
github tenstorrent org - these are all the teams prefixed with
`forge-developers-mlir-`. This PR changes CODEOWNERS to use teams
instead of personal aliases, where possible.

This change will make it easier to track which groups of people are
needed to review PRs.

Let's keep a working list of teams here so that folks can edit - when
we're ready, we'll then update the teams in github and mark this PR as
ready. Comment on the CODEOWNERS file as well where needed, and please
let me know if we need to add/remove some teams, and @staylorTT will
happily jump in to lend a hand :)

List of teams:
- forge-developers-mlir-build - @nsmithtt @mtopalovicTT @vwellsTT
- forge-developers-mlir-core - @sdjordjevicTT @svuckovicTT @mtopalovicTT
@jserbedzijaTT @azecevicTT
- forge-developers-mlir-d2m - @nsmithtt @vroubtsovTT @phizalev-TT 
- forge-developers-mlir-devops - @vmilosevic @jmcgrathTT @nsumrakTT
- forge-developers-mlir-explorer - @tapspatel @odjuricicTT
@vprajapati-tt @vcanicTT
- forge-developers-mlir-multidevice -
- forge-developers-mlir-optimizer - @odjuricicTT @rpavlovicTT
@sdjordjevicTT
- forge-developers-mlir-runtime - @jnie-TT @kmabeeTT @AleksKnezevic
@pilkicTT
- forge-developers-mlir-stablehlo - @mrakitaTT @mmanzoorTT @wooseokTT
@azecevicTT
- forge-developers-mlir-tosa - @mrakitaTT @sgligorijevicTT @sdjukicTT
@azecevicTT

### Checklist
- [ ] New/Existing tests provide coverage for changes