Add finish barrier after any enqueue_buffer_read ops to ensure sync (#3297)
### Ticket
NA

### Problem description
When transferring a buffer from device to host, we ought to barrier to
ensure queued commands complete before moving the data.

### What's changed
Whenever we generate an enqueue_buffer_read op, we also enqueue a finish
op to barrier right afterwards. In the future, we should insert a
targeted wait op instead and only await whichever ops we strictly depend
on, but this seems like an easy first step.

### Checklist
- [x] New/Existing tests provide coverage for changes