Add option to disable OpModel tests (#3281)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/2905

### Problem description
Tests that require opmodel last very long because they have to open a
device. Currently, they will run if the build flag
`TTMLIR_ENABLE_OPMODEL` is set.

### What's changed
Added one more level of granularity with `TTMLIR_ENABLE_OPMODEL_TESTS`
build flag. Rationale: If `TTMLIR_ENABLE_OPMODEL` is OFF whole part of
the project isn't built, so changes in the dialect API won't be caught,
if it's set to ON then `check-ttmlir` will last forever, so it's
suboptimal for incremental development if you're not working on things
related to optimizer. The flag is turned ON by default, so nothing
changes in the existing workflows; it just gives an opportunity to
developers to disable it if they don't need it.

### Checklist
- [ ] New/Existing tests provide coverage for changes