Moved PyKernel to tools dir + updated GH Action (#3206)
### Ticket
- PR closes #3111 

### Problem description
- Want to preserve `python` directory to hold the pure python bindings
related to `ttmlir`, tools and the rest will be maintained as `tools`
that rely on the compiler, but not necessarily core components of the
python bindings themselves. I consider PyKernel to be one of these tools
that leverages the compiler but will be exposed and packaged outwards as
a tool.

### What's changed
- Moved `python/pykernel` to `tools/pykernel`, added some CMake config
to support this.
- Updated GH Action to hopefully build the wheel from the correct
locations `/tools/pykernel`