Build-and-test workflow's logic fix (#3338)
### Problem description
The "Download build artifacts" job step had a poorly defined "if" clause
which resulted in the step always getting triggered.

### What's changed
Upon fixing the condition and it working as intended by the clause, all
of the tests failed ([workflow
preview](https://github.com/tenstorrent/tt-mlir/actions/runs/14991074320))
because they actually require the step to get executed.
In the end I just removed the condition so that it is obvious that we
always want it to run.
Workflow's behaviour hasn't changed.

### Tests
Tests are still passing
[workflow](https://github.com/tenstorrent/tt-mlir/actions/runs/14993789564)

---------

Co-authored-by: vvukomanTT <vvukomanTT@tenstorrent.com>