Compile and run ttnn-standalone in the CI (#3341)
### Ticket
#3248 

### Problem description
There is no test that compiles and runs ttnn-standalone in the CI, this
results in the tool often breaking with no visibility.

### What's changed
Test which builds, installs and runs ttnn-standalone has been added to
the `build-and-test` workflow.
The `build-tt-mlir-action` has been extended to publish the
`third_party` artifact because ttnn-standalone needs it for the build.

These changes should result in ttnn-standalone breaking less often.

### Checklist
- [ ] New/Existing tests provide coverage for changes

#### Note
This PR contains changes from [this
one](https://github.com/tenstorrent/tt-mlir/pull/3282).
The linked PR is for the actual fix for the tool and this one is for the
changes needed to compile and run it in the CI.

---------

Co-authored-by: svuckovic <svuckovic@tenstorrent.com>
Co-authored-by: vvukomanTT <vvukomanTT@tenstorrent.com>