Add more testcases with shapes not a multiple of tilesize (#3413)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3410

### Problem description
The recent uplift brought in a tt-metal commit that fixes an execution
issue with shapes whose dimensions are not multiples of the tile size.
To validate this fix, we added TTIR-builder test cases and run them in
CI to ensure the improvement is preserved going forward.

### What's changed
 - Added new TTIR-builder tests for n300 and llmbox.
- Marked several tests as run_error or fails_golden because of
additional issues that still need to be addressed (see #3410).

### Checklist
- [x] New/Existing tests provide coverage for changes