Skip connection check when querying system desc (#3425)
Checking device connection is done by querying the active ethernet cores
and the devices they are connected to. On blackhole chips, the
connection might be down, causing a runtime exception. As a temporary
workaround, we skip this step until an API is made public from the metal
side or when we need to support multiple BH devices.