[Fusing] Move conv2d weights to Host/RM workaround (#3465)
Context

When testing fusing for resnet we hit OOM when pushing inputs from host
to DRAM. This is because layouting logic prefers inputs in tile/dram
when possible. This wasn't the case before fusing because we had special
casing in layouting for function inputs which are used by conv2d to put
them in host/rm.
When fusing came some operations were commuted upward which made this
special cases useless in sense that now all inputs were forced in
tile/dram. Since resnet has weights which are not channel last when
tilized these tensors become quite large and they can't fit into dram.

Workaround

Long term we want to have all layouts in host/rm but this is not quite
simple. While trying to convert all to RM/Host we found some bugs which
are not easy fix (see
https://github.com/tenstorrent/tt-mlir/issues/3432). While we work
separately on this bugfixes we need some temporary solution for resnet
case. There were couple of solutions which were considered (if
interested I can add them to description separately), but in the end we
decided to go with discardable attributes and tag all weights of conv2d
before fusing and change layouting logic to check this attributes.

Longterm

Like said longterm fix is to move everything Host/RM
https://github.com/tenstorrent/tt-mlir/issues/3432 and let optimizer
change layouts when needed.