TTNN EmitC fixes for FEs bringup (#3435)
### Ticket
N/A

### Problem description
Few problems that I've spotted during the bringup of TTNN EmitC test
infrastructure in `tt-torch`.

### What's changed
- The `tt-torch` generates an entry function with the name `main`. This
causes an issue during compilation, as the `main` function is treated
specially in C/C++. I've special cased it, such that if MLIR function is
named `main`, I've renamed it to `_main`.
- Moved the whole logic of getting a mangled name of a function to the
`dylib.cpp`. There is no reason for FEs to know about mangling at all,
this is a temporary solution until the proper one is implemented
https://github.com/tenstorrent/tt-mlir/issues/2478.
- Few problems related to the installation of `TTNNStandalone`
component. Made `TTNNCompileSo` a library.

### Checklist
- [ ] New/Existing tests provide coverage for changes