Refactor GetDevice handling in EmitC dylib signature conversion (#3484)
### Ticket

https://github.com/tenstorrent/tt-mlir/issues/3397#issuecomment-2893207925

### Problem description
EmitC currently has special handling for getDevice ops which changes
function signatures, which was breaking logic in const-eval for
subgraphs which use getDevice ops.

### What's changed
Instead of passing the device as an argument to functions with
getDevices op, we want to utilize the existing conversion to get a
singleton device object. This requires adding a global setter for this
field in the precompiled .hpp as well, and invoking it via the dylib in
`ttrt` prior to execution.

### Checklist
- [X] New/Existing tests provide coverage for changes