Skip const-eval funcs in general ttrt run (#3578)
### Problem description
When const-eval hoists subgraphs, `ttrt run flatbuffer.ttnn` will
execute the individual subgraphs as well as the main function. This is
not ideal because it breaks ttrt perf, and it is confusing/opaque to any
user that such "programs" have been created.

### What's changed
Add a public flag to `Program`--all hoisted funcs will have it set to
false and be skipped in ttrt. Also fix the const-eval CI tests
(currently point to a dir that doesn't have tests/exist).

### Checklist
- [X] New/Existing tests provide coverage for changes