Fix bug with getInstance logic for emitC standalone (#3596)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3595

### Problem description
My recent changes to the way emitC handles getInstance seems to have
broken some tt_metal dtor ordering due to changing the type of static
variable involved. Unfortunately, CI did not catch this issue, but it is
causing segfaults when running ttnn-standalone.

### What's changed
This should resolve the issue by keeping the "ownedInstance" as a local
static variable instead of a class-level variable, as before. Locally,
this fixes the issue

### Checklist
- [ ] New/Existing tests provide coverage for changes--CI fix which
should have caught this are in the works by @vvukomanovicTT I believe