Introduce `ttkernel.fill_tile` op and lowering to EmitC (#3612)
`ttkenrel.fill_tile` op is added that allows setting a given DST
register tile to a specific f32 value.

### Problem description
Currently, there is no way to populate the DST tile with a specific f32
value.

### What's changed
`fill_tile(dst_reg_tile_index, val)` utility function that is part of
the compute API can be used to zero out a given DST tile with a specific
f32 value. This commit introduces a necessary op in `ttkernel` dialect
and also adds lowering of this op to `EmitC` dialect.

### Checklist
- [done] New/Existing tests provide coverage for changes