D2M preliminary multicore support (#3426)
The core of this change is twofold:
- Move all DMA lowering to a common place, i.e. GenericLowerDMAs. There
is now a new form of DMA, but this affords us simplicity in TTKernel
conversion + having all the DMA lowering related code in one place.
- Support for reblocking. Eventually we want to have these reblocking
affine maps encoded directly in the ViewOp from the beginning, but there
this support is blocked on this refactor #3389.

There are also a number of other cleanups and changes including:
- CoreIndexOp is not actually ConstantLike, I misunderstood this
attribute, it previously could be replaced with a constant.
- GenericOp support for single region datamovement thread type.
- TTKernel DMA lowering cleanup. Though mcast hasn't been touched yet.
- Test framework improvements.