Bringup erf op (#3617)
### Ticket
#3616 
### Problem description
As part of the effort to fuse the approximation for `gelu` to
`ttir.gelu`, I am bringing up the `erf` op as the decomposition of
`gelu` includes `erf`. `erf` is decomposed to an approximation. Future
PR for fusing that.

### What's changed
Adding erf op to the ttir/ttnn dialects - including support for lowering
through to flatbuffer.

### Checklist
- [x] New/Existing tests provide coverage for changes