Make tanh use accuracy mode and add a workaround to perform it in BF16 (#3546)
### Ticket
https://github.com/tenstorrent/tt-xla/issues/435 on tt-xla

### Problem description
Tanh was conclusively proven to be affecting accuracy of some models,
for example Albert

### What's changed
Runtime was changed always to pass the accuracy flag to tanh.
A workaround to run tanh in BF16 was added as otherwise setting accuracy
to true would throw an exception in the runtime. Accuracy in BF16 with
`accuracy=True` is still much better than in FP32 without it.

### Checklist
- [X] New/Existing tests provide coverage for changes