Fixing issue with toHost() function in runtime when called on multi device host tensors (#3629)
Currently, when calling the `toHost()` function (which calls
`::ttnn::distributed::get_device_tensors()`) on a multi device host
tensor, it will return a set of shardings, which also have a
`MULTI_DEVICE_HOST` storage type (even though they are logically on one
device each). This causes an issue if we try to make a new multidevice
tensor out of these shards, as multi device tensor creation from multi
device shards is not supported.
This PR fixes that (thanks @omilyutin-tt for the help!).