Improving TTIR dialect operation descriptions (#3529)
### Ticket
Part of https://github.com/tenstorrent/tt-mlir/issues/401

### Problem description
We have a bunch of rudimentary descriptions of our ops in the TTIR
dialect.

### What's changed
Generated descriptions for TTIR ops in a unified way. Most of the diff
is generated using Cluade, and I double checked a lot of ops, but errors
are possible.

### Checklist
- [x] New/Existing tests provide coverage for changes