Improved flatbuffer schema verification
The current flatbuffer schema verification check is flawed since it only
checks the git tag version is the same, this is problematic for 2
reasons:
1. False positives: You could have changed the schema as part of _this_
   commit and the compiler would be perfectly fine loading this system
   desc and parsing fields out (incorrectly).
2. False negatives: You could have committed a change that has nothing
   to do with flatbuffer schema, but this commit bumps the patch number
   which causes a system desc version mismatch.

This patch resolves both of these issues by hashing the flatbuffer schemas
and storing the hash in the flatbuffer.  This enables both the runtime
and the compiler to verify the schema is at the expected version.  It
has the added benefit of passing the verification when no flatbuffer
schema has changed.