Adding an `ttnn::AllGatherOp` workaround  (#3726)
Currently, TTNN does not support AllGather when the gather is happening
on one of the last two dimensions of the input tensor
(https://github.com/tenstorrent/tt-metal/issues/22654). This PR adds a
workaround to possibly reshape and permute the input tensor. The
workaround will change the input tensors in two cases:
1. If the input tensor has less than 3 dimensions, it will reshape the
input
tensor to have 3 dimensions by inserting a leading dimension of size 1,
perform the allGather operation, and then reshape the result back to the
  original shape.
2. If the allGatherDim is not the first dimension, it will permute the
input
  tensor to bring the allGatherDim to the front, perform the AllGather
  operation, and then permute the result back to the original order.