Fix consteval in canonical emitc path (#3716)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3594

### Problem description
Consteval was supported in dylib path used for testing emitc, but not in
the canonical path that a user would interact with, through
ttnn-standalone.

### What's changed
Edited `ttnn-modify-signatures-for-dylib` pass so that it works for both
the dylib and standalone paths, and renamed to `ttnn-tuplify-tensors`.
Locally tested that consteval works on benchmark resnet.

### Checklist
- [ ] New/Existing tests provide coverage for changes
TODO: Add consteval ttnn-standalone CI test subsequently, tracked here:
#3715