Fix metal runtime segfault
The root cause for this issue was that we were not allocating large
enough input tensors in the ttmlir runtime to be EnqueueWriteBuffer API
conformant.  This API expects the region size pointed to by the void*
argument to be aligned up to the shard spec's page size.

The compiler will always pick a page size equal to a 32x32 tile of the
respective dataformat in bytes.  This is an implicit contract between
the runtime and compiler and will probably have to remain that way.  In
the future we'll want to support zero copy from front-end frameworks and
will have to be allocation / alignment decisions before we ever have a
compiled graph.

The fix is to align up allocations in the runtime during toLayout to
tile size which guarantees that it's safe to pass into
EnqueueRead/WriteBuffer.

Additional changes:
- Cleanup of the TensorDesc to addionally track alignment plus
convenience constructors.
- Asan support for the runtime + metal runtime.  Tips on how to invoke
ttrt with asan:
```bash
# A ton of leaks are reported from metal, silence those with:
export ASAN_OPTIONS=detect_leaks=0
# Since ttrt is a python script, we need to preload the asan dylib as it
# doesn't support lazy loading.
export LD_PRELOAD=$(clang-17 -print-file-name=libclang_rt.asan-x86_64.so)
ttrt run ...
```