Dialect renaming: tt → ttcore (#3772)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3761

### Problem description
We have a clash with the Triton dialect, which uses the same tt prefix
for its dialect as our tt dialect.

### What's changed
Renamed the tt dialect to ttcore (now using the ttcore. prefix) across
the entire codebase, including all related tests. Most of the changes
are mechanical (search and replace), but I plan to do another thorough
review to ensure accuracy. If anyone notices anything that looks off or
inconsistent, please don’t hesitate to flag it.

Keeping the tt namespace as it already has a lot of changes, will update
the namespaces in a separate review.

### Checklist
- [x] New/Existing tests provide coverage for changes