Support multicore enqueue write buffer (#3830)
This change also comes with a fix to the page size calculation, namely
that we can decouple page size programming between circular buffers and
shard specs.

This commit implements the above by adding support for the pageSize
argument to getMemrefSizeBytes so that the shard spec can override it
accordingly. It also adds default alignment behavior for dram memory
space.