Add runtime api entry points to create runtime tensors from buffers containing unsupported data types (#3814)
Added an `enum` to represent data types that are unsupported.
- int64
- uint64
- float64
- int16
- int8
- bool

Added utility function which returns the data type we should cast to.
- `getUnsupportedDataTypeAlias`
- This is based off of `toTTMLIRSupportedDataType` which is used in the
`ElementTypeNormalization` pass.

Added runtime api for creating tensor from a buffer that includes data
in an unsupported data type.
- `createOwnedHostTensorFromUnsupportedDataType`
- This will assume that the host buffer containing the data holds an
unsupported data type and will cast all values inside to a supported
data type (determined by `getUnsupportedDataTypeAlias`) and print a
warning to the user that this is happening.
- I cannot create a version of this function which creates a borrowed
tensor as we cannot borrow a buffer containing an unsupported data type
without changing the values within the host buffer itself - which the
host assumes will not happen.

Added runtime api for copying a runtime tensor's values into a host
buffer but with an unsupported data format
- `copyIntoDestWithUnsupportedDataType` 
- This will cast each value in the runtime tensor to that which is
described by the `unsupportedDataType` param and copy the casted values
into the host buffer

**Why are we doing this?**

Until now we could simply ask the user to cast all model
parameters/constants to a supported data type. I am working on enabling
a flow which uses `torch-xla` to compile PyTorch models. `torch-xla`
will push python scalars to the device as `float64` or `int64` depending
on if it is an integer or float. If such values are used in a model then
PJRT will be asked to push this value to device as a 64-bit value, I
have found no way to stop this. This will happen even if the python
scalar is an op attribute (i.e `torch.arange` max and min values).

I am adding the functionality to create runtime tensors like this to
tt-mlir rather than tt-xla so other frontends may use it. For now, the
responsibility of keeping track of what the host expects the data type
to be for a given tensor will fall on the frontends (unless we decide to
implement that in tt-mlir now).