[Optimizer] Enable getOpConstraint Queries with Partially Specified MemoryConfigs (#3827)
### Ticket
#3242
#3408

### Problem description
In TTNN it is possible to create a `MemoryConfig` that is L1 and
sharded, but has an unspecified `shard_spec`. This means the tensor is
sharded but the physical shard shape and grid are unspecified. Such
`MemoryConfig`s cannot actually be used to allocate tensors. However,
some ops such as `matmul` accept them as `output_memory_config` so
callers can request a sharded output tensor while allowing the op to
decide the actual shard layout.

The optimizer uses `TTNNLayoutAttr` to model the layout of TTNN tensors.
This attribute requires a `GridAttr` and immediately computes the
physical shard shape. There is no mechanism to model the unspecified
`shardspec`. There is also no mechanism for the optimizer to signal to
the OpModel to only partially translate the `TTNNLayoutAttr`. This
blocks the proper exploration of sharding for `matmul` in the optimizer.

### What's changed
- Modify the `TTNNLayoutAttr` to add a flag, `ignorePhysicalLayout`.
When this flag is set, opModel skips the creation of a `shard_spec` when
translating the `TTNNLayoutAttr` to a `MemoryConfig`
- This flag does **not** bypass the need to provide a valid `GridAttr`.
Nor does it modify the any behaviour or property of the attribute. It is
a flag that must be read by the users of this Attr
- This flag defaults to **false**. All existing call sites for
`TTNNLayoutAttr::get` will call the `::get` method that sets this to
false
  - Added `::withIgnorePhysicalLayout()` and new constructor
- Added unit tests for correct conversion, and for calling
`getOpConstraints` on `matmul` with `ignorePhysicalLayout` set to true
for the `output_memory_config`
- Added an assertion to ensure `tensor_spec`s are not created from
layouts with this flag set to true
- Closes #3242
- Closes #3408

### Rationale 
There are two alternatives to this approach:
1. Make the `GridAttr` an optional parameter and bypass its usage in
`TTNNLayoutAttr`
- Problem: not having a `GridAttr` means we need to also null the
`AffineMap` and the `Memref`. At that point, almost every method of
`TTNNLayoutAttr` would have to return null. Every call site would need
to handle the potential nullability or risk a segfault
2. Introduce a new attribute `TTNNPartialLayoutAttr`
- Problem: need to define easy transformations to and from
`TTNNLayoutAttr`. Lots of API churn to accept this in the opModel

I think the current approach minimizes churn and avoids the pitfalls of
the above alternatives. The worst case is that a `TTNNLayoutAttr` is
created with a default meaningless `GridAttr` and `ignorePhysicalLayout
== true` but a user does not respect that. I think garbage data is
better than potential segfaults.

### Checklist
- [X] New/Existing tests provide coverage for changes